restructured the class to make it generic for all APIs. Now Query class
