+++ project/views.py
@@ -1,23 +1,22 @@
 """Views for Tidbits."""
 
 import logging
-import tweepy
 
 from flask import redirect, render_template, session, url_for
 
 from project.app import app
 from project.forms import QueryForm
-from project.twitter import TwitterQuery
+from project.query import Query
 
 
 @app.route('/', methods=['GET', 'POST'])
 def home():
     """Home page."""
-    twitter = TwitterQuery()
+    # query_instance = Query()
     form = QueryForm()
     if form.validate_on_submit():
         session['query_result'] = form.query.data
-        print twitter.search(session['query_result'])
+        # print query_instance.twitter_search(session['query_result'])
         logging.debug('User queried:|%s|', session['query_result'])
         return redirect(url_for('.results'))
     return render_template('index.html', title='Search', form=form)
